crud commentaire avec jointure user

+++ src/Controller/CommentaireController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Commentaire;
+use App\Entity\User;
 use App\Form\CommentaireType;
 use App\Repository\CommentaireRepository;
 use Doctrine\ORM\EntityManagerInterface;
@@ -23,22 +24,35 @@ final class CommentaireController extends AbstractController
     }
 
     #[Route('/new', name: 'app_commentaire_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, CommentaireRepository $commentaireRepository): Response
     {
         $commentaire = new Commentaire();
         $form = $this->createForm(CommentaireType::class, $commentaire);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $user = $this->getUser();
+            if ($user instanceof User) {
+                $commentaire->setUser($user);
+            }
+
             $entityManager->persist($commentaire);
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_commentaire_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_commentaire_new', [], Response::HTTP_SEE_OTHER);
+        }
+
+        $commentaires = $commentaireRepository->findAll();
+        $editForms = [];
+        foreach ($commentaires as $comment) {
+            $editForms[$comment->getId()] = $this->createForm(CommentaireType::class, $comment)->createView();
         }
 
         return $this->render('commentaire/new.html.twig', [
             'commentaire' => $commentaire,
-            'form' => $form,
+            'form' => $form->createView(),
+            'commentaires' => $commentaires,
+            'edit_forms' => $editForms,
         ]);
     }
 
@@ -50,38 +64,37 @@ final class CommentaireController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_commentaire_edit', methods: ['GET', 'POST'])]
+    #[Route('/{id}/edit', name: 'app_commentaire_edit', methods: ['POST'])]
     public function edit(Request $request, Commentaire $commentaire, EntityManagerInterface $entityManager): Response
     {
+        if ($commentaire->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('You cannot edit this comment.');
+        }
+
         $form = $this->createForm(CommentaireType::class, $commentaire);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_commentaire_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_commentaire_new', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('commentaire/edit.html.twig', [
-            'commentaire' => $commentaire,
-            'form' => $form,
-        ]);
+        return $this->redirectToRoute('app_commentaire_new', [], Response::HTTP_SEE_OTHER);
     }
 
     #[Route('/{id}', name: 'app_commentaire_delete', methods: ['POST'])]
     public function delete(Request $request, Commentaire $commentaire, EntityManagerInterface $entityManager): Response
     {
+        if ($commentaire->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('You cannot delete this comment.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$commentaire->getId(), $request->request->get('_token'))) {
             $entityManager->remove($commentaire);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('app_commentaire_index', [], Response::HTTP_SEE_OTHER);
-    }
-
-    #[Route('/forum', name: 'app_forum')]
-    public function showForum(): Response
-    {
-        return $this->render('commentaire/show.html.twig');
+        return $this->redirectToRoute('app_commentaire_new', [], Response::HTTP_SEE_OTHER);
     }
 }
